// src/components/forms/fieldset-cart/fieldset-promo/enum.ts
enum TagNames {
  FIELDSET = 'fieldset',
  BUTTON = 'button',
  DIV = 'div',
  LABEL = 'label',
  INPUT = 'input',
}

enum Styles {
  FIELDSET = 'fieldset-promo',
  FIELDSET_SHOW = 'show',
  LEGEND = 'legend',
  BUTTONS_CONTAINER = 'promo-buttons-container',
  BUTTON_CANCEL = 'cancel-button',
  BUTTON_SAVE = 'save-button',
  BUTTONS_HIDE = 'hide',
  BUTTONS_SHOW = 'show',
  HIGHLIGHT = 'highlight',
  INPUT_CONTAINER = 'input-container',
  LABEL = 'label',
  INPUT = 'input-promo',
  INPUT_ERROR = 'input--error',
}

enum Events {
  INPUT = 'input',
}

enum Attributes {
  ID = 'id',
  ID_VALUE_INPUT = 'email',
  TYPE = 'type',
  FOR = 'for',
  NAME = 'name',
  NAME_VALUE = 'email',
  PLACEHOLDER = 'placeholder',
  PLACEHOLDER_VALUE = 'Enter email',
}

enum Contents {
  LABEL = 'Email',
}

export { TagNames, Styles, Events, Contents, Attributes };